refactor: remove conditions from repository and move them to CustomerService

+++ src/main/java/com/vodafone/service/CustomerService.java
@@ -53,16 +53,26 @@ public class CustomerService {
             throw new NullIdException("Null customer id is provided");
         if (updatedCustomer == null)
             throw new NullCustomerException("Null updated customer entity is provided");
+        if (get(id) == null)
+            throw new NullCustomerException("Customer not found with provided id");
         return customerRepository.update(id, updatedCustomer);
     }
 
     public boolean updateStatusActivated(String email) {
+        if (email == null)
+            throw new NullPointerException("Null email is provided");
+        if (getByMail(email) == null) {
+            throw new NullCustomerException("Customer not found with email provided");
+        }
         return customerRepository.updateStatusActivated(email);
     }
 
     public boolean delete(Long id) {
         if (id == null)
             throw new NullIdException("Null customer id is provided");
+        if (get(id) == null) {
+            throw new NullCustomerException("Customer not found with id provided");
+        }
         return customerRepository.delete(id);
     }
 
@@ -93,12 +103,18 @@ public class CustomerService {
             throw new NullPointerException("Null Email is provided");
         if (password == null)
             throw new NullPointerException("Null Password is provided");
+        if (getByMail(email) == null) {
+            throw new NullCustomerException("Customer not found with provided email");
+        }
         return customerRepository.resetPassword(email, password);
     }
 
     public boolean expireOtp(String userName) {
         if (userName == null)
             throw new NullPointerException("Null Username is provided");
+        if (getByUserName(userName) == null) {
+            throw new NullCustomerException("Customer not found with provided username");
+        }
         return this.customerRepository.expireOtp(userName);
     }
